+++ lib/src/balance/api_balance_service.dart
@@ -1,5 +1,8 @@
-import 'package:http/http.dart' show Response;
+import 'dart:convert';
 
+import 'package:messagebird_dart/src/balance/model/balance.dart';
+
+import '../../src/general/model/api_error.dart';
 import '../base_service.dart';
 import 'balance_service.dart';
 
@@ -10,5 +13,13 @@ class ApiBalanceService extends BaseService implements BalanceService {
       : super(accessKey, timeout: timeout, features: features);
 
   @override
-  Future<Response> read() => get('/balance');
+  Future<Balance> read() async {
+    final response = await get('/balance');
+    if (response.statusCode == 200) {
+      return Future.value(Balance.fromJson(json.decode(response.body)));
+    } else {
+      final ApiError error = ApiError.fromJson(json.decode(response.body));
+      throw Exception(error.code);
+    }
+  }
 }