+++ lib/src/balance/balance_service.dart
@@ -1,8 +1,8 @@
-import 'package:http/http.dart' show Response;
+import 'package:messagebird_dart/src/balance/model/balance.dart';
 
 /// Balance service interface.
 // ignore: one_member_abstracts
 abstract class BalanceService {
   /// Get account balance.
-  Future<Response> read();
+  Future<Balance> read();
 }