+++ CHANGELOG.md
@@ -1,2 +1,6 @@
+## 0.2.0
+- Completed all endpoint implementations
+- Initial model (balance, callflows & calls)
+
 ## 0.1.0
 - Initial version